LayoutTests/imported/w3c/web-platform-tests/dom/lists/DOMTokenList-value-expected.txt


FAIL DOMTokenList value assert_equals: value should return the literal value expected "   a  a b " but got "a b"
